refactor: centralize lenis instance
