--- my-corp-site/src/scripts/gsap.js
@@ -1,5 +1,0 @@
-import { gsap } from "gsap";
-import ScrollTrigger from "gsap/ScrollTrigger";
-
-
-gsap.registerPlugin(ScrollTrigger);
